create table-component module refactor elements for tab bar
