+++ ts/component/table-components.ts
@@ -1,0 +1,48 @@
+import { siteConfig } from "../util/site-config";
+import { urls } from "../util/urls";
+import { FilterDefinition, SortDefinition } from "./table-builder";
+
+export const latestActivitySort: SortDefinition = {
+  columnTitle: "Latest Activity",
+  descending: true,
+  type: "date",
+};
+
+export var caseFilters: FilterDefinition[] = [
+  {
+    title: "Assay",
+    key: "ASSAY",
+    type: "text",
+    autocompleteUrl: urls.rest.autocomplete.assayNames,
+  },
+  {
+    title: "Donor",
+    key: "DONOR",
+    type: "text",
+    autocompleteUrl: urls.rest.autocomplete.donorNames,
+  },
+  {
+    title: "Pending",
+    key: "PENDING",
+    type: "dropdown",
+    values: siteConfig.pendingStates,
+  },
+  {
+    title: "Pipeline",
+    key: "PIPELINE",
+    type: "dropdown",
+    values: siteConfig.pipelines,
+  },
+  {
+    title: "Project",
+    key: "PROJECT",
+    type: "text",
+    autocompleteUrl: urls.rest.autocomplete.projectNames,
+  },
+  {
+    title: "Requisition",
+    key: "REQUISITION",
+    type: "text",
+    autocompleteUrl: urls.rest.autocomplete.requisitionNames,
+  },
+];
